Fix a few issues regarding invalid thread accesses and canceling scan jobs' behavior
